Ajout de trois nouvelles routes

+++ src/main/java/com/jee/crud/repository/CommandRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommandRepository extends JpaRepository<Command, Long> {
+public interface CommandRepository extends JpaRepository<Command, Integer> {
     List<Command> findByPaymentStatusCommand(PaymentStatusCommand paymentStatusCommand);
 }
